+++ src/error.ts
@@ -1,0 +1,6 @@
+export class URLParamError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "URLParamError";
+  }
+}
